requirements.txt
dash>=2.14.0
pandas>=1.5.0
gspread>=5.7.0
google-auth>=2.15.0
openpyxl>=3.0.0
gunicorn>=20.1.0